move MetaTagsParser to common module
